// src/app/front/services/twitter-prop.enum.ts
export enum TwitterProp {
  TITLE = 'twitter:title',
  CARD = 'twitter:card',
  CREATOR = 'twitter:creator',
  SITE = 'twitter:site',
  DESCRIPTION = 'twitter:description',
  IMAGE = 'twitter:image'
}

export enum TwitterCardType {
  SUMMARY = 'summary',
  SUMMARY_LARGE_IMAGE = 'summary_large_image',
  APP = 'app',
  PLAYER = 'player'
}